pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal-rag"
version = "0.1.0"
description = "Multimodal RAG system with custom cross-modal retriever"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.36.0",
    "Pillow>=10.0.0",
    "PyMuPDF>=1.23.0",
    "ultralytics>=8.0.0",
    "qdrant-client>=1.7.0",
    "google-generativeai>=0.3.0",
    "gradio>=4.0.0",
    "omegaconf>=2.3.0",
    "datasets>=2.16.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ipykernel>=6.27.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
